Fix solaris build
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@947238 13f79535-47bb-0310-9956-ffa450edef68

+++ proxy/signals.cc
@@ -303,7 +303,7 @@ signal_handler(int sig, siginfo_t * t, void *c)
     if (t->si_code <= 0) {
 #if (HOST_OS == solaris)
       snprintf(sig_msg, sizeof(sig_msg), "NOTE: Traffic Server received User Sig %d from pid: %d uid: %d\n",
-               sig, (int)t->si_pid, t->si_uid);
+               sig, (int)t->si_pid, (int)t->si_uid);
 #else
       snprintf(sig_msg, sizeof(sig_msg), "NOTE: Traffic Server received User Sig %d from pid: %d uid: %d\n",
                sig, t->si_pid, t->si_uid);
